+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/DigDurabilityEnchantmentMixin.java
@@ -1,19 +1,24 @@
 package insane96mcp.iguanatweaksreborn.mixin;
 
+import com.llamalad7.mixinextras.injector.ModifyReturnValue;
 import insane96mcp.iguanatweaksreborn.module.experience.enchantments.EnchantmentsFeature;
-import net.minecraft.util.RandomSource;
-import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.enchantment.DigDurabilityEnchantment;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(DigDurabilityEnchantment.class)
 public abstract class DigDurabilityEnchantmentMixin {
-	@Inject(at = @At(value = "RETURN"), method = "shouldIgnoreDurabilityDrop", cancellable = true)
-	private static void onChanceToIgnoreDurabilityDrop(ItemStack stack, int lvl, RandomSource random, CallbackInfoReturnable<Boolean> cir) {
+	@ModifyReturnValue(at = @At(value = "RETURN"), method = "shouldIgnoreDurabilityDrop")
+	private static boolean onChanceToIgnoreDurabilityDrop(boolean original) {
 		if (EnchantmentsFeature.isUnbreakingOverhaul())
-			cir.setReturnValue(false);
+			return false;
+		return original;
+	}
+
+	@ModifyReturnValue(at = @At(value = "RETURN"), method = "getMaxLevel")
+	private int onChanceToIgnoreDurabilityDrop(int original) {
+		if (EnchantmentsFeature.isUnbreakingOverhaul())
+			return 5;
+		return original;
 	}
 }