+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/PlayerMixin.java
@@ -93,7 +93,7 @@ public abstract class PlayerMixin extends LivingEntity {
 	private float iguanatweaksreborn$applyBonusEnchantmentsEfficiency(float efficiency, BlockState state, @Nullable BlockPos pos, @Local(ordinal = 0) int efficiencyLvl) {
 		//Remove vanilla efficiency
 		efficiency -= (float)(efficiencyLvl * efficiencyLvl + 1);
-		return ISOEventFactory.getBonusEnchantmentEfficiency((Player) (Object) this, state, this.getMainHandItem(), efficiency);
+		return ISOEventFactory.getEfficiencyWithEnchantments((Player) (Object) this, state, this.getMainHandItem(), efficiency);
 	}
 
 	/**