+++ src/main/java/insane96mcp/iguanatweaksreborn/module/items/misc/MiscItem.java
@@ -3,6 +3,7 @@ package insane96mcp.iguanatweaksreborn.module.items.misc;
 import insane96mcp.iguanatweaksreborn.data.generator.ISOItemTagsProvider;
 import insane96mcp.iguanatweaksreborn.event.ISOEventFactory;
 import insane96mcp.iguanatweaksreborn.module.Modules;
+import insane96mcp.insanelib.InsaneLib;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.LoadFeature;
@@ -60,8 +61,8 @@ public class MiscItem extends Feature {
 
 		ItemStack stack = event.getItemStack();
 		if (stack.getItem() instanceof DiggerItem diggerItem){
-			float toolEfficiency = ISOEventFactory.getBonusEnchantmentEfficiency(event.getEntity(), null, stack, diggerItem.speed);
-			event.getToolTip().add(CommonComponents.space().append(Component.translatable(TOOL_MINING_SPEED_LANG, ItemStack.ATTRIBUTE_MODIFIER_FORMAT.format(toolEfficiency))).withStyle(ChatFormatting.DARK_GREEN));
+			float toolEfficiency = ISOEventFactory.getEfficiencyWithEnchantments(event.getEntity(), null, stack, diggerItem.speed);
+			event.getToolTip().add(CommonComponents.space().append(Component.translatable(TOOL_MINING_SPEED_LANG, InsaneLib.ONE_DECIMAL_FORMATTER.format(toolEfficiency))).withStyle(ChatFormatting.DARK_GREEN));
 		}
 	}